requirements.txt
streamlit>=1.30.0
pandas>=1.5.0
python-dotenv>=1.0.0
pandasai>=1.5.10
duckdb>=0.9.1
# Groq API request
requests>=2.31.0
openpyxl>=3.1.2  # for reading Excel files
matplotlib>=3.7.0